finish insert app

+++ Models/Model/TsneCoordinate.cs
@@ -7,8 +7,8 @@ namespace WorldBT.Models.Model
     {
         public Guid Id { get; set; }
         public Guid PatientId { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public decimal X { get; set; }
+        public decimal Y { get; set; }
 
         public virtual Patient Patient { get; set; }
     }
